% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_futime.R
\name{calc_futime}
\alias{calc_futime}
\title{Calculate follow-up time per case until end of follow-up depending on pat_status}
\usage{
calc_futime(
  wide_df,
  futime_var_new = "p_futimeyrs",
  fu_end,
  dattype = "zfkd",
  check = TRUE,
  time_unit = "years",
  status_var = "p_status",
  lifedat_var = NULL,
  fcdat_var = NULL,
  spcdat_var = NULL
)
}
\arguments{
\item{wide_df}{dataframe in wide format}

\item{futime_var_new}{Name of the newly calculated variable for follow-up time. Default is p_futimeyrs.}

\item{fu_end}{end of follow-up in time format YYYY-MM-DD.}

\item{dattype}{Type of cancer registry data. Can be "seer" or "zfkd". Default is "zfkd".}

\item{check}{Check newly calculated variable p_status by printing frequency table. Default is TRUE.}

\item{time_unit}{Unit of follow-up time (can be "days", "weeks", "months", "years"). Default is "years".}

\item{status_var}{Name of the patient status variable that was previously created. Default is p_status.}

\item{lifedat_var}{Name of variable containing Date of Death. Will override dattype preset.}

\item{fcdat_var}{Name of variable containing Date of Primary Cancer diagnosis. Will override dattype preset.}

\item{spcdat_var}{Name of variable containing Date of SPC diagnosis Will override dattype preset.}
}
\value{
wide_df
}
\description{
Calculate follow-up time per case until end of follow-up depending on pat_status
}
