% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_futime_tt.R
\name{calc_futime_tt}
\alias{calc_futime_tt}
\title{Calculate follow-up time per case until end of follow-up depending on pat_status - tidytable version}
\usage{
calc_futime_tt(
  wide_df,
  futime_var_new = "p_futimeyrs",
  fu_end,
  dattype = NULL,
  check = TRUE,
  time_unit = "years",
  status_var = "p_status",
  lifedat_var = NULL,
  fcdat_var = NULL,
  spcdat_var = NULL,
  quiet = FALSE
)
}
\arguments{
\item{wide_df}{dataframe or data.table in wide format}

\item{futime_var_new}{Name of the newly calculated variable for follow-up time. Default is p_futimeyrs.}

\item{fu_end}{end of follow-up in time format YYYY-MM-DD.}

\item{dattype}{can be "zfkd" or "seer" or NULL. Will set default variable names if dattype is "seer" or "zfkd". Default is NULL.}

\item{check}{Check newly calculated variable p_futimeyrs by printing frequency table. Default is TRUE.}

\item{time_unit}{Unit of follow-up time (can be "days", "weeks", "months", "years"). Default is "years".}

\item{status_var}{Name of the patient status variable that was previously created. Default is p_status.}

\item{lifedat_var}{Name of variable containing Date of Death. Will override dattype preset.}

\item{fcdat_var}{Name of variable containing Date of Primary Cancer diagnosis. Will override dattype preset.}

\item{spcdat_var}{Name of variable containing Date of SPC diagnosis Will override dattype preset.}

\item{quiet}{If TRUE, warnings and messages will be suppressed. Default is FALSE.}
}
\value{
wide_df
}
\description{
Calculate follow-up time per case until end of follow-up depending on pat_status - tidytable version
}
