% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_crosstab.R
\name{ir_crosstab}
\alias{ir_crosstab}
\title{Calculate crude incidence rates and crosstabulate results by break variables}
\usage{
ir_crosstab(
  df,
  dattype = "zfkd",
  count_var,
  xbreak_var = "none",
  ybreak_vars,
  collapse_ci = FALSE,
  add_total = "no",
  add_n_percentages = FALSE,
  futime_var = NULL,
  alpha = 0.05
)
}
\arguments{
\item{df}{dataframe in wide format}

\item{dattype}{can be "zfkd" or "seer" or empty. Will set default variable names from dataset.}

\item{count_var}{variable to be counted as observed case. Should be 1 for case to be counted.}

\item{xbreak_var}{variable from df by which rates should be stratified in columns of result df. Default is "none".}

\item{ybreak_vars}{variables from df by which rates should be stratified in rows of result df. Multiple variables will result in
appended rows in result df. y_break_vars is required.}

\item{collapse_ci}{If TRUE upper and lower confidence interval will be collapsed into one column separated by "-". Default is FALSE.}

\item{add_total}{option to add a row of totals. Can be either "no" for not adding such a row or "top" or "bottom" for adding it at the first or last row. Default is "no".}

\item{add_n_percentages}{option to add a column of percentages for n_base in its respective yvar_group. Can only be used when xbreak_var = "none". Default is FALSE.}

\item{futime_var}{variable in df that contains follow-up time per person (in years). Default is set if dattype is given.}

\item{alpha}{significance level for confidence interval calculations. Default is alpha = 0.05 which will give 95 percent confidence intervals.}
}
\value{
df
}
\description{
Calculate crude incidence rates and crosstabulate results by break variables
}
