% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_crosstab_byfutime.R
\name{ir_crosstab_byfutime}
\alias{ir_crosstab_byfutime}
\title{Calculate crude incidence rates and crosstabulate results by break variables (Copy of ir_crosstab for the special case of cumulative FU-times as xbreak_var)}
\usage{
ir_crosstab_byfutime(df, dattype = "zfkd", count_var,
  futime_breaks = c(6, 12, 60, 120), ybreak_vars, collapse_ci = FALSE,
  add_total = "no", futime_var = NULL, alpha = 0.05)
}
\arguments{
\item{df}{dataframe in wide format}

\item{dattype}{can be "zfkd" or "seer" or empty. Will set default variable names from dataset.}

\item{count_var}{variable to be counted as observed case. Should be 1 for case to be counted.}

\item{futime_breaks}{vector that indicates split points for follow-up time groups (in months) that will be used as xbreak_var.
Default is c(6, 12, 60, 120) that will result in 5 groups (up to 6 months, 6-12 months, 1-5 years, 5-10 years, 10+ years)}

\item{ybreak_vars}{variables from df by which rates should be stratified in rows of result df. Multiple variables will result in
appended rows in result df. y_break_vars is required.}

\item{collapse_ci}{If TRUE upper and lower confidence interval will be collapsed into one column separated by "-". Default is FALSE.}

\item{add_total}{option to add a row of totals. Can bei either "no" for not adding such a row or "top" or "bottom" for adding it at the first or last row. Default is "no".}

\item{futime_var}{variable in df that contains follow-up time per person (in years). Default is set if dattype is given.}

\item{alpha}{signifcance level for confidence interval calculations. Default is alpha = 0.05 which will give 95 percent confidence intervals.}
}
\value{
df
}
\description{
Calculate crude incidence rates and crosstabulate results by break variables (Copy of ir_crosstab for the special case of cumulative FU-times as xbreak_var)
}
