% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msSPChelpR-package.R
\docType{package}
\name{msSPChelpR-package}
\alias{msSPChelpR-package}
\title{msSPChelpR: Helper Functions for Second Primary Cancer Analyses}
\description{
A collection of helper functions for analyzing Second Primary Cancer data, including functions to reshape data, to calculate patient states and analyze cancer incidence.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marianschmidt.github.io/msSPChelpR/}
  \item Report bugs at \url{https://github.com/marianschmidt/msSPChelpR/issues}
}

}
\author{
\strong{Maintainer}: Marian Eberl \email{marian.eberl@tum.de} (\href{https://orcid.org/0000-0001-6584-3197}{ORCID})

}
\keyword{internal}
