% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msSPChelpR.R
\docType{package}
\name{msSPChelpR}
\alias{msSPChelpR}
\alias{msSPChelpR-package}
\title{msSPChelpR}
\description{
R helper functions for Second Primary Cancer analyses
}
\section{Main functions}{


Functions include asir, ir_crosstabl, pat_status, reshape.
}

\section{Package context}{

This package follows the principles of the "tidyverse".
}

\keyword{internal}
