% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_status_dtplyr.R
\name{pat_status_dtplyr}
\alias{pat_status_dtplyr}
\title{Calculate patient status at specific end of follow-up using dtplyr}
\usage{
pat_status_dtplyr(
  df,
  fu_end = NULL,
  dattype = "zfkd",
  status_var = "p_status",
  life_var = NULL,
  spc_var = NULL,
  birthdat_var = NULL,
  lifedat_var = NULL,
  fcdat_var = NULL,
  spcdat_var = NULL,
  life_stat_alive = NULL,
  life_stat_dead = NULL,
  spc_stat_yes = NULL,
  spc_stat_no = NULL,
  lifedat_fu_end = NULL,
  check = TRUE,
  as_labelled_factor = FALSE
)
}
\arguments{
\item{df}{dataframe in wide format}

\item{fu_end}{end of follow-up in time format YYYY-MM-DD.}

\item{dattype}{Type of cancer registry data. Can be "seer" or "zfkd". Default is "zfkd".}

\item{status_var}{Name of the newly calculated variable for patient status. Default is p_status.}

\item{life_var}{Name of variable containing life status. Will override dattype preset.}

\item{spc_var}{Name of variable containing SPC status. Will override dattype preset.}

\item{birthdat_var}{Name of variable containing Date of Birth. Will override dattype preset.}

\item{lifedat_var}{Name of variable containing Date of Death. Will override dattype preset.}

\item{fcdat_var}{Name of variable containing Date of Primary Cancer diagnosis. Will override dattype preset.}

\item{spcdat_var}{Name of variable containing Date of SPC diagnosis Will override dattype preset.}

\item{life_stat_alive}{Value for alive status in life_var. Will override dattype preset.}

\item{life_stat_dead}{Value for dead status in life_var. Will override dattype preset.}

\item{spc_stat_yes}{Value for SPC occured in spc_var. Will override dattype preset.}

\item{spc_stat_no}{Value for no SPC in spc_var. Will override dattype preset.}

\item{lifedat_fu_end}{Date of last FU of alive status in registry data. Will override dattype preset (2017-03-31 for zfkd; 2018-12-31 for seer).}

\item{check}{Check newly calculated variable p_status}

\item{as_labelled_factor}{If true, output status_var as labelled factor variable. Default is FALSE.}
}
\value{
df
}
\description{
Calculate patient status at specific end of follow-up using dtplyr
}
