% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_us}
\alias{population_us}
\title{US Populations Data}
\format{
A data frame with the following variables:
\describe{
\item{\code{region}}{Region / Registry}
\item{\code{year}}{Year group}
\item{\code{sex}}{Sex}
\item{\code{age}}{Age group}
\item{\code{race}}{Race}
\item{\code{population_pyar}}{Population Years used for rate calculation (PYAR)}
\item{\code{population_n_per_year}}{Absolute Population in single years or periods (PYAR / 5 years)]}
}
}
\usage{
population_us
}
\description{
Dataset that contains different standard populations needed to run some package functions
}
\keyword{datasets}
