% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renumber_time_id.R
\name{renumber_time_id}
\alias{renumber_time_id}
\title{Renumber the time ID per case (i.e. Tumor sequence)}
\usage{
renumber_time_id(
  df,
  new_time_id_var,
  dattype = "zfkd",
  case_id_var = NULL,
  time_id_var = NULL
)
}
\arguments{
\item{df}{dataframe}

\item{new_time_id_var}{Name of the newly calculated variable for time_id. Required.}

\item{dattype}{Type of cancer registry data. Can be "seer" or "zfkd". Default is "zfkd".}

\item{case_id_var}{String or vector of strings with name of ID variable indicating same patient.
E.g. \code{idvar="PUBCSNUM"} for SEER data.}

\item{time_id_var}{String with name of variable that indicates diagnosis per patient.
E.g. \code{timevar="SEQ_NUM"} for SEER data.}
}
\value{
df
}
\description{
Renumber the time ID per case (i.e. Tumor sequence)
}
