% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renumber_time_id_dt.R
\name{renumber_time_id_dt}
\alias{renumber_time_id_dt}
\title{Renumber the time ID per case (i.e. Tumor sequence) using data.table}
\usage{
renumber_time_id_dt(
  df,
  new_time_id_var,
  dattype = "zfkd",
  case_id_var = NULL,
  time_id_var = NULL,
  timevar_max = Inf
)
}
\arguments{
\item{df}{dataframe}

\item{new_time_id_var}{Name of the newly calculated variable for time_id. Required.}

\item{dattype}{Type of cancer registry data. Can be "seer" or "zfkd". Default is "zfkd".}

\item{case_id_var}{String or vector of strings with name of ID variable indicating same patient.
E.g. \code{idvar="PUBCSNUM"} for SEER data.}

\item{time_id_var}{String with name of variable that indicates diagnosis per patient.
E.g. \code{timevar="SEQ_NUM"} for SEER data.}

\item{timevar_max}{Numeric; default Inf. Maximum number of cases per id.
All tumors > timevar_max will be deleted before renumbering}
}
\value{
df
}
\description{
Renumber the time ID per case (i.e. Tumor sequence) using data.table
}
