% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_wide.R
\name{reshape_wide}
\alias{reshape_wide}
\title{Reshape dataset to wide format}
\usage{
reshape_wide(df, timevar, idvar, chunks = 10)
}
\arguments{
\item{df}{dataframe}

\item{timevar}{String with name of variable that indicates diagnosis per patient.
E.g. \code{timevar="SEQ_NUM"} for SEER data.}

\item{idvar}{String or vector of strings with name of ID variable indicating same patient.
E.g. \code{idvar="PUBCSNUM"} for SEER data.}

\item{chunks}{Numeric; default 10.}
}
\value{
df
}
\description{
Reshape dataset to wide format
}
