% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_wide_tidyr.R
\name{reshape_wide_tidyr}
\alias{reshape_wide_tidyr}
\title{Reshape dataset to wide format}
\usage{
reshape_wide_tidyr(df, case_id_var, time_id_var, timevar_max = 6,
  var_selection = c("_all"))
}
\arguments{
\item{df}{dataframe}

\item{case_id_var}{String or vector of strings with name of ID variable indicating same patient.
E.g. \code{idvar="PUBCSNUM"} for SEER data.}

\item{time_id_var}{String with name of variable that indicates diagnosis per patient.
E.g. \code{timevar="SEQ_NUM"} for SEER data.}

\item{timevar_max}{Numeric; default 6. Maximum number of cases per id. 
All tumors > timevar_max will be deleted before reshaping.}

\item{var_selection}{Vector of variables to keep. Default is "_all".}
}
\value{
df
}
\description{
Reshape dataset to wide format
}
