% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir}
\alias{sir}
\title{Calculate standardized incidence ratios with costum grouping variables}
\usage{
sir(
  df,
  dattype = "zfkd",
  expcount_src = "refrates",
  xbreak_vars = "none",
  ybreak_vars = "none",
  summarize_groups = "none",
  count_var,
  refrates_df = rates,
  add_total = "no",
  collapse_ci = FALSE,
  stdpop_df = standard_population,
  refpop_df = population,
  std_pop = "ESP2013",
  region_var = NULL,
  agegroup_var = NULL,
  sex_var = NULL,
  year_var = NULL,
  icdcat_var = NULL,
  futime_var = NULL,
  pyar_var = NULL,
  alpha = 0.05
)
}
\arguments{
\item{df}{dataframe in wide format}

\item{dattype}{can be "zfkd" or "seer" or empty. Will set default variable names from dataset.}

\item{expcount_src}{can bei either "refrates" or "cohort"}

\item{xbreak_vars}{variables from df by which rates should be stratified in columns of result df. Default is "none".}

\item{ybreak_vars}{variables from df by which rates should be stratified in rows of result df. Multiple variables will result in
appended rows in result df. Default is "none".}

\item{summarize_groups}{option to define summarizing stratified groups. Default is "none". 
If you want to define variables that should be summarized into one group, you can chose from region_var, sex_var, year_var. 
Define multiple summarize variables by summarize_groups = c("region", "sex", "year")}

\item{count_var}{variable to be counted as observed case. Should be 1 for case to be counted.}

\item{refrates_df}{df where reference rate from general population are defined. It is assumed that refrates_df has the columns 
"region" for region, "sex" for gender, "age" for age-groups (can be single ages or 5-year brackets), "year" for time period (can be single year or 5-year brackets), 
"incidence_crude_rate" for incidence rate in the respective age/gender/year cohort.
refrates_df must use the same category coding of age, sex, region, year and icdcat as agegroup_var, sex_var, region_var, year_var and icdcat_var.}

\item{add_total}{option to add a row of totals. Can bei either "no" for not adding such a row or "top" or "bottom" for adding it at the first or last row. Default is "no".}

\item{collapse_ci}{If TRUE upper and lower confidence interval will be collapsed into one column separated by "-". Default is FALSE.}

\item{stdpop_df}{df where standard population is defined. It is assumed that stdpop_df has the columns "sex" for gender, "age" for age-groups,
"standard_pop" for name of standard population (e.g. "European Standard Population 2013) and "population_n" for size of standard population age-group.
stdpop_df must use the same category coding of age and sex as agegroup_var and sex_var.}

\item{refpop_df}{df where reference population data is defined. Only required if option futime = "refpop" is chosen. It is assumed that refpop_df has the columns 
"region" for region, "sex" for gender, "age" for age-groups (can be single ages or 5-year brackets), "year" for time period (can be single year or 5-year brackets), 
"population_pyar" for person-years at risk in the respective age/gender/year cohort.
refpop_df must use the same category coding of age, sex, region, year and icdcat as agegroup_var, sex_var, region_var, year_var and icdcat_var.}

\item{std_pop}{can be either "ESP2013, ESP1976, WHO1960. Only applies to expcount_src = "cohort".}

\item{region_var}{variable in df that contains information on region where case was incident. Default is set if dattype is given.}

\item{agegroup_var}{variable in df that contains information on age-group. Default is set if dattype is given.}

\item{sex_var}{variable in df that contains information on gender. Default is set if dattype is given.}

\item{year_var}{variable in df that contains information on year or year-period when case was incident. Default is set if dattype is given.}

\item{icdcat_var}{variable in df that contains information on ICD code of case diagnosis. Default is set if dattype is given.}

\item{futime_var}{variable in df that contains follow-up time per person (in years) in cohort (can only be used with futime_src = "cohort"). Default is set if dattype is given.}

\item{pyar_var}{variable in refpop_df that contains person-years-at-risk in reference population (can only be used with futime_src = "refpop") Default is set if dattype is given.}

\item{alpha}{signifcance level for confidence interval calculations. Default is alpha = 0.05 which will give 95 percent confidence intervals.}
}
\value{
df
}
\description{
Calculate standardized incidence ratios with costum grouping variables
}
