% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_population}
\alias{standard_population}
\title{Standard Populations}
\format{
A data frame with the following variables:
\describe{
\item{\code{standard_pop}}{Standard Population}
\item{\code{sex}}{Sex}
\item{\code{age}}{Age group}
\item{\code{population_n}}{Absolute Population number in standard population age group}
\item{\code{group_proportion}}{Proportion of age-group in gender-specific total population}
}
}
\usage{
standard_population
}
\description{
Dataset that contains different standard populations needed to run some package functions
}
\keyword{datasets}
