% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sir_results.R
\name{summarize_sir_results}
\alias{summarize_sir_results}
\title{Summarize detailed SIR results}
\usage{
summarize_sir_results(
  sir_df,
  summarize_groups,
  summarize_icdcat,
  output = "long",
  add_total_row = "no",
  add_total_fu = "no",
  collapse_ci = FALSE,
  fubreak_var_name = "fu_time",
  ybreak_var_name = "yvar_name",
  xbreak_var_name = "xvar_name",
  alpha = 0.05
)
}
\arguments{
\item{sir_df}{dataframe with stratified sir results created using the sir or sir_byfutime functions}

\item{summarize_groups}{option to define summarizing stratified groups. Default is "none". 
If you want to define variables that should be summarized into one group, you can chose from region_var, sex_var, year_var. 
Define multiple summarize variables by summarize_groups = c("region", "sex", "year")}

\item{summarize_icdcat}{option to summarize over all ICD codes}

\item{output}{Define the format of the output. Can be either "nested" for nested dataframe with fubreak_var and xbreak_var in separate sub_tables (purrr).
Or "wide" for wide format where fubreak_var and xbreak_var are appended as columns.
Or "long" for long format where sir_df is not reshaped, but just summarized (ybreak_var, xbreak_var and fubreak_var remain in rows).
Default is "long".}

\item{add_total_row}{option to add a row of totals. Can bei either "no" for not adding such a row or "top" or "bottom" for adding it at the first or last row. Default is "no".}

\item{add_total_fu}{option to add totals for follow-up time. Can bei either "no" for not adding such a column or "left" or "right" for adding it at the first or last column. Default is "no".}

\item{collapse_ci}{If TRUE upper and lower confidence interval will be collapsed into one column separated by "-". Default is FALSE.}

\item{fubreak_var_name}{Name of variable with futime stratification. Default is "fu_time".}

\item{ybreak_var_name}{Name of variable with futime stratification. Default is "yvar_name".}

\item{xbreak_var_name}{Name of variable with futime stratification. Default is "xvar_name".}

\item{alpha}{signifcance level for confidence interval calculations. Default is alpha = 0.05 which will give 95 percent confidence intervals.}
}
\description{
Summarize detailed SIR results
}
