% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_refrates_icd2}
\alias{us_refrates_icd2}
\title{US Reference Rates for Cancer using ICD-O 2digit code for cancer site}
\format{
A data frame with the following variables:
\describe{
\item{\code{t_site}}{Tumor Site}
\item{\code{region}}{Region / Region groups}
\item{\code{year}}{Year / Periods}
\item{\code{sex}}{Sex}
\item{\code{age}}{Age / Age groups}
\item{\code{race}}{Race}
\item{\code{comment}}{Comment}
\item{\code{incidence_cases}}{Incident Cases (raw count)}
\item{\code{incidence_crude_rate}}{Incidence Rate (crude rate)}
\item{\code{population_pyar}}{Population Years used for rate calculation (PYAR)}
\item{\code{population_n_per_year}}{Absolute Population number used for rate calculation (PYAR / 5 years)}
}
}
\usage{
us_refrates_icd2
}
\description{
Synthetic dataset of reference incidence rates for the US population to demonstrate package functions
}
\keyword{datasets}
