% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_status.R
\name{vital_status}
\alias{vital_status}
\title{Calculate vital status at end of follow-up depending on pat_status}
\usage{
vital_status(
  df,
  status_var = "p_status",
  life_var_new = "p_alive",
  check = TRUE
)
}
\arguments{
\item{df}{dataframe in wide format}

\item{status_var}{Name of the patient status variable that was previously created. Default is p_status.}

\item{life_var_new}{Name of the newly calculated variable for patient vital status. Default is p_alive.}

\item{check}{Check newly calculated variable p_status by printing frequency table. Default is TRUE.}
}
\value{
df
}
\description{
Calculate vital status at end of follow-up depending on pat_status
}
