% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msSPChelpR.R
\name{msSPChelpR}
\alias{msSPChelpR}
\title{msSPChelpR}
\description{
R helper functions for Second Primary Cancer analyses
}
\section{Main functions}{


Functions include asir, ir_crosstab, pat_status, reshape, sir_byfutime and summarize_sir_results.
}

\section{Package context}{

This package follows the principles of the "tidyverse".
}

\keyword{internal}
