% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_long.R
\name{reshape_long}
\alias{reshape_long}
\title{Reshape dataset to long format}
\usage{
reshape_long(df, case_id_var, time_id_var, chunks = 10, keep_vars = c("_all"))
}
\arguments{
\item{df}{dataframe}

\item{case_id_var}{String or vector of strings with name of ID variable indicating same patient.
E.g. \code{idvar="PUBCSNUM"} for SEER data.}

\item{time_id_var}{String with name of variable that indicates diagnosis per patient.
E.g. \code{timevar="SEQ_NUM"} for SEER data.}

\item{chunks}{Numeric; default 10.}

\item{keep_vars}{Vector of variables to keep. Default is "_all".}
}
\value{
df
}
\description{
Reshape dataset to long format
}
