% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_long_tidyr.R
\name{reshape_long_tidyr}
\alias{reshape_long_tidyr}
\title{Reshape dataset to long format}
\usage{
reshape_long_tidyr(df, case_id_var, time_id_var, time_id_num = TRUE,
  constant_vars = NULL, drop_na = FALSE, var_selection = c("_all"),
  names_pattern = "(.*)\\\\.(.*)")
}
\arguments{
\item{df}{dataframe}

\item{case_id_var}{String or vector of strings with name of ID variable indicating same patient.
E.g. \code{idvar="PUBCSNUM"} for SEER data.}

\item{time_id_var}{String with name of variable that indicates diagnosis per patient.
E.g. \code{timevar="SEQ_NUM"} for SEER data.}

\item{time_id_num}{Logical that indicates whether time_id_var should be transformed to numeric or not.
Default is \code{time_id_num=TRUE}.}

\item{constant_vars}{Vector of variables that are constant for all times per case_id, so these variables will just be filled during pivot_longer
Default is NULL.}

\item{drop_na}{Logical to indicate whether empty rows after pivoting should be dropped (activates option values_drop_na in tidyr::pivot_longer).
Default is drop_na = FALSE.}

\item{var_selection}{Vector of variables to keep after pivoting. Default is "_all".}

\item{names_pattern}{A regular expression containing matching groups (()). Default is names_pattern = "(.*)\\.(.*)" where variable name and time_id are separated by a dot ".".}
}
\value{
df
}
\description{
Reshape dataset to long format
}
