% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_long_tidyr.R
\name{reshape_long_tidyr}
\alias{reshape_long_tidyr}
\title{Reshape dataset to wide format - tidyr version}
\usage{
reshape_long_tidyr(wide_df, case_id_var, time_id_var, datsize = Inf)
}
\arguments{
\item{wide_df}{dataframe}

\item{case_id_var}{String with name of ID variable indicating same patient.
E.g. \code{idvar="PUBCSNUM"} for SEER data.}

\item{time_id_var}{String with name of variable that indicates diagnosis per patient.
E.g. \code{timevar="SEQ_NUM"} for SEER data.}

\item{datsize}{Number of rows to be taken from df. This parameter is mainly for testing. Default is Inf so that df is fully processed.}

\item{timevar_max}{Numeric; default 6. Maximum number of cases per id.
All tumors > timevar_max will be deleted before reshaping.}
}
\value{
long_df
}
\description{
Reshape dataset to wide format - tidyr version
}
