% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_wide.R
\name{reshape_wide}
\alias{reshape_wide}
\title{Reshape dataset to wide format}
\usage{
reshape_wide(df, case_id_var, time_id_var, chunks = 10, timevar_max = 6)
}
\arguments{
\item{df}{dataframe}

\item{case_id_var}{String or vector of strings with name of ID variable indicating same patient.
E.g. \code{idvar="PUBCSNUM"} for SEER data.}

\item{time_id_var}{String with name of variable that indicates diagnosis per patient.
E.g. \code{timevar="SEQ_NUM"} for SEER data.}

\item{chunks}{Numeric; default 10.}

\item{timevar_max}{Numeric; default 6. Maximum number of cases per id.
All tumors > timevar_max will be deleted before reshaping.}
}
\value{
df
}
\description{
Reshape dataset to wide format
}
